add juice.rb,suica.rb
